requirements.txt
Django==5.1.7
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
openai==1.12.0
tensorflow==2.16.1
numpy==1.26.3 